requirements.txt
python-gsmmodem>=0.9
pyserial>=3.5